+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 2016-02-14
+- adding window.performance.timing support
+    - adding domContentLoaded
+    - adding onLoad
+
 ## 2016-02-13
 - update loading engine
 - rename the example files